rag_sql_qa_1.1.1/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.6
langchain-community>=0.0.6
chromadb>=0.4.0
openai>=1.0.0
python-dotenv
